// app/containers/Projects/types.ts
export interface IProjectPermission {
  downloadPermission: boolean
  schedulePermission: number
  sharePermission: boolean
  sourcePermission: number
  viewPermission: number
  vizPermission: number
  widgetPermission: number
}

export interface IProject {
  createBy?: { avatar?: string; id?: number; username?: string; email: string }
  permission?: IProjectPermission
  initialOrgId?: number
  userId: number
  isStar?: boolean
  isTransfer?: boolean
  isFavorites?: boolean
  type?: string
  name?: string
  id?: number
  description?: string
  pic?: string
  orgId?: number
  visibility?: boolean
  starNum?: number
  createTime?: string
  updateTime?: string
  isMinePro?: boolean
}
// ...
export interface IProjectFormFieldProps {
  id?: number
  orgId_hc?: string
  orgId?: number
  name?: string
  description?: string
  visibility?: string
  pic?: string
}

export interface IProjectsFormProps {
  type?: string
  organizations?: any
  onTransfer?: () => any
  onModalOk?: () => any
  modalLoading: boolean
  currentPro: Partial<IProject>
  onCheckUniqueName?: (
    pathname: any,
    data: any,
    resolve: () => any,
    reject: (error: string) => any
  ) => any
}
// ...
export enum FormType {
  'transfer' = '移交',
  'add' = '创建',
  'edit' = '修改'
}